% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BF_.R
\name{BF_binomial}
\alias{BF_binomial}
\title{Bayes Factors for Binomial point null hypotheses039114}
\usage{
BF_binomial(x, success, null_par, hyper_par = c(1, 1), in_favour = "H0")
}
\arguments{
\item{x}{Atomic vector of the type integer, double or character.}

\item{success}{Number or character indicating which level of \code{x} corresponds to a success.}

\item{null_par}{The parameter of the Binomial distribution under the null hipothesis.}

\item{a, }{b Shape parameters of the Beta prior. Both must be greater than zero}
}
\description{
Computes Bayes factors in favour of a point null hypothesis in the context of a Binomial statistical model with a Beta prior.
}
\examples{

}
