% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BF_.R
\name{BF_multinomial}
\alias{BF_multinomial}
\title{Bayes Factors for Multinomial point null hypotheses}
\usage{
BF_multinomial(x, categories, null_par, hyper_par = 1, in_favour = "H0")
}
\description{
Computes Bayes factors in favour of a point null hypothesis in the context of a Multinomial statistical model with a Dirichlet prior.
}
