% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfactor_.R
\name{bfactor_multinomial}
\alias{bfactor_multinomial}
\title{Bayes Factors for Multinomial point null hypotheses}
\usage{
bfactor_multinomial(x, categories, null_par, hyper_par = 1, transf = "level")
}
\description{
Computes Bayes factors in favour of a point null hypothesis in the context of a Multinomial statistical model with a Dirichlet prior.
}
