% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfactor_.R
\name{bfactor_to_prob}
\alias{bfactor_to_prob}
\title{Use Bayes factors to Update Prior Probabilities to Posterior Probabilities}
\usage{
bfactor_to_prob(bf, pi_null = 0.5)
}
\arguments{
\item{bf}{a Bayes factor}

\item{pi_null}{The prior probability of the null hypothesis.}
}
\description{
Use Bayes factors to Update Prior Probabilities to Posterior Probabilities
}
