% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classical-tests.R
\name{chisq_test_multinomial}
\alias{chisq_test_multinomial}
\title{Pearson's Chi-Squared Test for Multinomial Proportions}
\usage{
chisq_test_multinomial(x, categories, null_par, simulate.p.value = FALSE)
}
\description{
Pearson's Chi-Squared Test for Multinomial Proportions
}
