% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbeta.R
\name{mbeta}
\alias{mbeta}
\title{Multivariate Beta Function}
\usage{
mbeta(x)
}
\arguments{
\item{x}{A numeric vector of non-negative values.}
}
\value{
If all elements of \code{x} are greater than zero then \code{mbeta} returns a numeric vector of \code{\link[base]{length}} 1. If any element of \code{x} is equal to zero then \code{mbeta} returns \code{inf}.
}
\description{
\loadmathjax Underflow-robust implementation of the multivariate Beta function.
}
\details{
Consider the numeric vector \mjeqn{x = (x_1, ..., x_k)}{x = (x_1, ..., x_k)} with \mjeqn{x_i > 0 \, \forall i \in \lbrace 1, \dots, k \rbrace}{x_i > 0 for all i in \{1, ..., k \}}. The multivariate Beta function is defined as:

\mjdeqn{B(x) = \prod_{i=1}^{k} \Gamma (x_i) \Big/ \Gamma \left(\sum_{i=1}^{k} x_i \right)}{B(x) = product_{i=1}^{k} Gamma(x_i) / Gamma(sum_{i=1}^{k} x_i)}

where \mjeqn{\Gamma}{Gamma} is the Gamma function. To avoid underflow, \code{mbeta} implements an alternative representation of the Beta function. Taking the natural logarithm of both sides in the latter equation and then exponentiating we obtain:

\mjdeqn{\operatorname{B}(x) = \operatorname{exp} \,  \left\lbrace \sum_{i=1}^{k}{\operatorname{log} \, \Gamma(x_i)} - \operatorname{log} \, \Gamma \left(\sum_{i = 1}^{k}x_i\right)\right\rbrace}{B(x) = exp(sum_\{i=1\}^\{k\} logGamma(x_i) - logGamma(sum_\{i = 1\}^\{k\} x_i))}

With this formula we can take advantage of the \code{\link[base]{lgamma}} function, which is itself underflow-robust. Note that \mjeqn{\operatorname{B}(x)}{B(x)} is only defined in \mjeqn{\Re}{R} if all elements of \code{x} are greater than zero and is infinite if any element of \code{x} is equal to zero \insertCite{becker1972}{polya}.
}
\examples{
mbeta(c(1, 1)) # same as beta(1,1)
mbeta(c(1, 1, 1))
mbeta(c(1, 2, 4))
mbeta(rep(0.5, 10))
mbeta(1:5)

}
\references{
\insertAllCited{}
}
\seealso{
\itemize{
\item \code{\link[polya]{lmbeta}} for an underflow-robust implementation of the natural logarithm of the Beta function.
}
}
