% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classical-tests.R
\name{nigrini_z_test}
\alias{nigrini_z_test}
\title{Nigrini's Z-test for Binomial Proportions with Continuity Correction}
\usage{
nigrini_z_test(x, success, null_par)
}
\description{
Nigrini's Z-test for Binomial Proportions with Continuity Correction
}
