% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary-functions.R
\name{smsdigit}
\alias{smsdigit}
\title{Returns the second most significant digit for each element in the input vector}
\usage{
smsdigit(x)
}
\description{
Returns the second most significant digit for each element in the input vector
}
