% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_.R
\name{test.null.binomial}
\alias{test.null.binomial}
\title{Bayes Factors, Strength of Evidence, Posterior probabilities and Chi-Squared test for Binomial Proportions}
\usage{
test.null.binomial(
  x,
  null.par,
  sucess,
  a = 1,
  b = 1,
  pi_null = 0.5,
  transf = "level",
  bf.round = 2,
  probs.round = 3
)
}
\description{
Bayes Factors, Strength of Evidence, Posterior probabilities and Chi-Squared test for Binomial Proportions
}
\examples{

test.null.binomial(austria_bl1, null.par = theta_benford(1)[1], sucess = 1, transf = "log10")
}
