% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_.R
\name{test.null.multinomial}
\alias{test.null.multinomial}
\title{teafdf}
\usage{
test.null.multinomial(
  x,
  categories,
  null_par,
  hyper_par = 1,
  haldane = FALSE,
  pi_null = 0.5,
  transf = "level",
  bf_round = 2,
  probs_round = 3
)
}
\description{
teafdf
}
\examples{

test.null.multinomial(austria_bl1, null_par = theta_benford(1), categories = 1:9, hyper_par=1, transf = "log10")
}
