% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_.R
\name{test.null.binomial}
\alias{test.null.binomial}
\title{Bayes Factors, Strength of Evidence, Posterior probabilities and Chi-Squared test for Binomial Proportions}
\usage{
test.null.binomial(
  x,
  success,
  null_par,
  hyper.par = c(1, 1),
  haldane = FALSE,
  pi_null = 0.5,
  transf = "level",
  bf_round = 2,
  probs_round = 3
)
}
\description{
Bayes Factors, Strength of Evidence, Posterior probabilities and Chi-Squared test for Binomial Proportions
}
\examples{

test.null.binomial(austria_bl1, success = 1, null_par = theta_benford(1)[1])
}
